# Make LFVM OpCodes binary compatible with Tosca definitions. 