--- a/go/interpreter/lfvm/converter_test.go
+++ b/go/interpreter/lfvm/converter_test.go
@@ -186,7 +186,7 @@ func TestConvert_AllValidOperationsAreCoveredByConversionTable(t *testing.T) {
 			continue
 		}
 
-		if op_2_op[code] == INVALID && vm.IsValid(code) {
+		if OpCode(code) == INVALID && vm.IsValid(code) {
 			t.Errorf("Missing instruction coverage for %v", code)
 		}
 	}
@@ -222,10 +222,10 @@ func TestConvertWithObserver_MapsEvmToLfvmPositions(t *testing.T) {
 	}
 
 	for _, p := range pairs {
-		in := vm.OpCode(code[p.evm])
-		want := op_2_op[in]
-		if vm.PUSH1 <= in && in <= vm.PUSH32 {
-			want = PUSH1 + OpCode(in-vm.PUSH1)
+		toscaOpCode := vm.OpCode(code[p.evm])
+		want := OpCode(toscaOpCode)
+		if vm.PUSH1 <= toscaOpCode && toscaOpCode <= vm.PUSH32 {
+			want = PUSH1 + OpCode(toscaOpCode-vm.PUSH1)
 		}
 		got := res[p.lfvm].opcode
 		if want != got {
@@ -251,10 +251,10 @@ func TestConvertWithObserver_PreservesJumpDestLocations(t *testing.T) {
 
 		// Check that all operations are mapped to matching operations.
 		for _, p := range pairs {
-			in := vm.OpCode(code[p.evm])
-			want := op_2_op[in]
-			if vm.PUSH1 <= in && in <= vm.PUSH32 {
-				want = PUSH1 + OpCode(in-vm.PUSH1)
+			toscaOpCode := vm.OpCode(code[p.evm])
+			want := OpCode(toscaOpCode)
+			if vm.PUSH1 <= toscaOpCode && toscaOpCode <= vm.PUSH32 {
+				want = PUSH1 + OpCode(toscaOpCode-vm.PUSH1)
 			}
 			got := res[p.lfvm].opcode
 			if want != got {
@@ -273,14 +273,22 @@ func TestConvertWithObserver_PreservesJumpDestLocations(t *testing.T) {
 	}
 }
 
-func BenchmarkConvertLongExampleCode(b *testing.B) {
-	converter, err := NewConverter(ConversionConfig{})
+func benchmarkConvertCode(b *testing.B, code []byte, config ConversionConfig) {
+	converter, err := NewConverter(config)
 	if err != nil {
 		b.Fatalf("failed to create converter: %v", err)
 	}
 	for i := 0; i < b.N; i++ {
-		converter.Convert(longExampleCode, nil)
+		converter.Convert(code, nil)
 	}
+}
+
+func BenchmarkConvertLongExampleCodeNoCache(b *testing.B) {
+	benchmarkConvertCode(b, longExampleCode, ConversionConfig{CacheSize: -1})
+}
+
+func BenchmarkConvertLongExampleCode(b *testing.B) {
+	benchmarkConvertCode(b, longExampleCode, ConversionConfig{})
 }
 
 func BenchmarkConversionCacheLookupSpeed(b *testing.B) {