--- a/go/interpreter/lfvm/instruction_test.go
+++ b/go/interpreter/lfvm/instruction_test.go
@@ -13,15 +13,27 @@ package lfvm
 import "testing"
 
 func TestInstruction_String(t *testing.T) {
-	instruction := Instruction{opcode: STOP, arg: 0x0000}
-	if got, want := instruction.String(), "STOP"; got != want {
-		t.Errorf("Instruction.String() = %q, want %q", got, want)
+
+	tests := []struct {
+		instruction Instruction
+		want        string
+	}{
+		{Instruction{opcode: STOP, arg: 0x0000}, "STOP"},
+		{Instruction{opcode: PUSH1, arg: 0x0001}, "PUSH1 0x0001"},
+		{Instruction{opcode: PUSH2, arg: 0x0002}, "PUSH2 0x0002"},
+		{Instruction{opcode: DATA, arg: 0x0002}, "DATA 0x0002"},
+		{Instruction{opcode: JUMP_TO, arg: 0x0002}, "JUMP_TO 0x0002"},
+		{Instruction{opcode: PUSH2_JUMP, arg: 0x0002}, "PUSH2_JUMP 0x0002"},
+		{Instruction{opcode: PUSH2_JUMPI, arg: 0x0002}, "PUSH2_JUMPI 0x0002"},
+		{Instruction{opcode: PUSH1_PUSH4_DUP3, arg: 0x0002}, "PUSH1_PUSH4_DUP3 0x0002"},
 	}
 
-	instruction = Instruction{opcode: PUSH1, arg: 0x0001}
-	if got, want := instruction.String(), "PUSH1 0x0001"; got != want {
-		t.Errorf("Instruction.String() = %q, want %q", got, want)
+	for _, tt := range tests {
+		if got := tt.instruction.String(); got != tt.want {
+			t.Errorf("Instruction.String() = %q, want %q", got, tt.want)
+		}
 	}
+
 }
 
 func TestCode_String(t *testing.T) {