--- a/go/interpreter/lfvm/stack_usage.go
+++ b/go/interpreter/lfvm/stack_usage.go
@@ -10,8 +10,6 @@
 
 package lfvm
 
-import "fmt"
-
 // stackUsage defines the combined effect of an instruction on the stack. Each
 // instruction is accessing a range of elements on the stack relative to the
 // stack pointer. The range is given by the interval [from, to) where from is
@@ -23,8 +21,8 @@ type stackUsage struct {
 
 // computeStackUsage computes the stack usage of the given opcode. The result
 // is a stackUsage struct that defines the combined effect of the instruction
-// on the stack. If the opcode is not supported, an error is returned.
-func computeStackUsage(op OpCode) (stackUsage, error) {
+// on the stack. If the opcode is not known, zero stack usage is reported.
+func computeStackUsage(op OpCode) stackUsage {
 
 	// For single instructions it is easiest to define the stack usage based on
 	// the opcode's pops and pushes.
@@ -38,49 +36,49 @@ func computeStackUsage(op OpCode) (stackUsage, error) {
 	}
 
 	if PUSH1 <= op && op <= PUSH32 {
-		return makeUsage(0, 1), nil
+		return makeUsage(0, 1)
 	}
 	if DUP1 <= op && op <= DUP16 {
-		return makeUsage(int(op-DUP1+1), int(op-DUP1+2)), nil
+		return makeUsage(int(op-DUP1+1), int(op-DUP1+2))
 	}
 	if SWAP1 <= op && op <= SWAP16 {
-		return makeUsage(int(op-SWAP1+2), int(op-SWAP1+2)), nil
+		return makeUsage(int(op-SWAP1+2), int(op-SWAP1+2))
 	}
 	if LOG0 <= op && op <= LOG4 {
-		return makeUsage(int(op-LOG0+2), 0), nil
+		return makeUsage(int(op-LOG0+2), 0)
 	}
 
 	switch op {
 	case JUMPDEST, JUMP_TO, STOP:
-		return makeUsage(0, 0), nil
+		return makeUsage(0, 0)
 	case PUSH0, MSIZE, ADDRESS, ORIGIN, CALLER, CALLVALUE, CALLDATASIZE,
 		CODESIZE, GASPRICE, COINBASE, TIMESTAMP, NUMBER,
 		PREVRANDAO, GASLIMIT, PC, GAS, RETURNDATASIZE,
 		SELFBALANCE, CHAINID, BASEFEE, BLOBBASEFEE:
-		return makeUsage(0, 1), nil
+		return makeUsage(0, 1)
 	case POP, JUMP, SELFDESTRUCT:
-		return makeUsage(1, 0), nil
+		return makeUsage(1, 0)
 	case ISZERO, NOT, BALANCE, CALLDATALOAD, EXTCODESIZE,
 		BLOCKHASH, MLOAD, SLOAD, TLOAD, EXTCODEHASH, BLOBHASH:
-		return makeUsage(1, 1), nil
+		return makeUsage(1, 1)
 	case MSTORE, MSTORE8, SSTORE, TSTORE, JUMPI, RETURN, REVERT:
-		return makeUsage(2, 0), nil
+		return makeUsage(2, 0)
 	case ADD, SUB, MUL, DIV, SDIV, MOD, SMOD, EXP, SIGNEXTEND,
 		SHA3, LT, GT, SLT, SGT, EQ, AND, XOR, OR, BYTE,
 		SHL, SHR, SAR:
-		return makeUsage(2, 1), nil
+		return makeUsage(2, 1)
 	case CALLDATACOPY, CODECOPY, RETURNDATACOPY, MCOPY:
-		return makeUsage(3, 0), nil
+		return makeUsage(3, 0)
 	case ADDMOD, MULMOD, CREATE:
-		return makeUsage(3, 1), nil
+		return makeUsage(3, 1)
 	case EXTCODECOPY:
-		return makeUsage(4, 0), nil
+		return makeUsage(4, 0)
 	case CREATE2:
-		return makeUsage(4, 1), nil
+		return makeUsage(4, 1)
 	case STATICCALL, DELEGATECALL:
-		return makeUsage(6, 1), nil
+		return makeUsage(6, 1)
 	case CALL, CALLCODE:
-		return makeUsage(7, 1), nil
+		return makeUsage(7, 1)
 	}
 
 	// For super-instructions, we need to decompose the instruction into its
@@ -88,16 +86,12 @@ func computeStackUsage(op OpCode) (stackUsage, error) {
 	if op.isSuperInstruction() {
 		usages := []stackUsage{}
 		for _, subOp := range op.decompose() {
-			usage, err := computeStackUsage(subOp)
-			if err != nil {
-				return stackUsage{}, err
-			}
-			usages = append(usages, usage)
+			usages = append(usages, computeStackUsage(subOp))
 		}
-		return combineStackUsage(usages...), nil
+		return combineStackUsage(usages...)
 	}
 
-	return stackUsage{}, fmt.Errorf("unsupported opcode: %v", op)
+	return stackUsage{}
 }
 
 // combineStackUsage combines the given stack usages into a single stack usage.