… go/interpreter/lfvm/interpreter_test.go
… go/interpreter/lfvm/opcode_test.go
… go/interpreter/lfvm/stack_usage_test.go