@LuisPH3 2024-09-10
Note, I was not happy with last night performance runs, I repeated the test and validated the commit hashes. 
- Left column corresponds to the head of the master branch at the time of the experiment
- Right column is the head of the branch for this PR.