@LuisPH3 2024-09-11
New benchmarks:
```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/integration_test/interpreter
cpu: Intel(R) Xeon(R) Platinum 8481C CPU @ 2.70GHz
                         │    main     │             opcodes#0              │             opcodes#1              │           opcodes_16b_addr            │
                         │   sec/op    │   sec/op     vs base               │   sec/op     vs base               │    sec/op      vs base                │
Inc/1/lfvm-8               1.674µ ± 0%   1.719µ ± 0%  +2.72% (p=0.000 n=20)   1.719µ ± 0%  +2.72% (p=0.000 n=20)    1.729µ ±  4%   +3.29% (p=0.000 n=20)
Inc/10/lfvm-8              1.671µ ± 0%   1.718µ ± 0%  +2.81% (p=0.000 n=20)   1.718µ ± 0%  +2.81% (p=0.000 n=20)    1.727µ ±  1%   +3.35% (p=0.000 n=20)
Fib/1/lfvm-8               1.434µ ± 0%   1.491µ ± 0%  +4.01% (p=0.000 n=20)   1.491µ ± 0%  +4.01% (p=0.000 n=20)    1.493µ ±  0%   +4.12% (p=0.000 n=20)
Fib/5/lfvm-8               9.085µ ± 1%   9.212µ ± 1%  +1.39% (p=0.000 n=20)   9.212µ ± 1%  +1.39% (p=0.000 n=20)    9.190µ ± 25%   +1.15% (p=0.000 n=20)
Fib/10/lfvm-8              96.36µ ± 2%   95.56µ ± 0%       ~ (p=0.597 n=20)   95.56µ ± 0%       ~ (p=0.597 n=20)   122.57µ ± 22%  +27.20% (p=0.005 n=20)
Fib/15/lfvm-8              1.047m ± 2%   1.049m ± 0%       ~ (p=0.086 n=20)   1.049m ± 0%       ~ (p=0.086 n=20)    1.054m ± 28%   +0.65% (p=0.001 n=20)
Fib/20/lfvm-8              11.80m ± 2%   11.59m ± 0%       ~ (p=0.102 n=20)   11.59m ± 0%       ~ (p=0.102 n=20)    11.66m ± 29%        ~ (p=0.134 n=20)
Sha3/1/lfvm-8              635.9n ± 0%   657.0n ± 1%  +3.32% (p=0.000 n=20)   657.0n ± 1%  +3.32% (p=0.000 n=20)    637.8n ±  0%   +0.31% (p=0.001 n=20)
Sha3/10/lfvm-8             1.941µ ± 0%   1.964µ ± 0%  +1.18% (p=0.000 n=20)   1.964µ ± 0%  +1.18% (p=0.000 n=20)    1.958µ ±  3%   +0.88% (p=0.000 n=20)
Sha3/100/lfvm-8            14.63µ ± 2%   14.57µ ± 0%  -0.40% (p=0.036 n=20)   14.57µ ± 0%  -0.40% (p=0.036 n=20)    19.09µ ± 23%  +30.48% (p=0.002 n=20)
Sha3/1000/lfvm-8           141.8µ ± 3%   138.0µ ± 0%       ~ (p=0.211 n=20)   138.0µ ± 0%       ~ (p=0.211 n=20)    139.9µ ± 32%        ~ (p=0.221 n=20)
Arith/1/lfvm-8             1.820µ ± 0%   1.903µ ± 0%  +4.56% (p=0.000 n=20)   1.903µ ± 0%  +4.56% (p=0.000 n=20)    1.896µ ±  5%   +4.15% (p=0.000 n=20)
Arith/10/lfvm-8            4.814µ ± 1%   5.006µ ± 0%  +3.99% (p=0.000 n=20)   5.006µ ± 0%  +3.99% (p=0.000 n=20)    4.930µ ± 18%   +2.41% (p=0.000 n=20)
Arith/100/lfvm-8           34.19µ ± 2%   35.43µ ± 0%  +3.62% (p=0.000 n=20)   35.43µ ± 0%  +3.62% (p=0.000 n=20)    53.17µ ± 35%  +55.53% (p=0.000 n=20)
Arith/280/lfvm-8           95.37µ ± 2%   98.20µ ± 0%  +2.97% (p=0.000 n=20)   98.20µ ± 0%  +2.97% (p=0.000 n=20)   146.74µ ± 35%  +53.87% (p=0.000 n=20)
Memory/1/lfvm-8            2.931µ ± 0%   3.067µ ± 0%  +4.66% (p=0.000 n=20)   3.067µ ± 0%  +4.66% (p=0.000 n=20)    3.029µ ±  3%   +3.34% (p=0.000 n=20)
Memory/10/lfvm-8           7.143µ ± 0%   7.440µ ± 1%  +4.16% (p=0.000 n=20)   7.440µ ± 1%  +4.16% (p=0.000 n=20)    7.257µ ± 23%   +1.59% (p=0.000 n=20)
Memory/100/lfvm-8          48.64µ ± 0%   50.54µ ± 1%  +3.91% (p=0.000 n=20)   50.54µ ± 1%  +3.91% (p=0.000 n=20)    51.02µ ± 33%   +4.90% (p=0.001 n=20)
Memory/1000/lfvm-8         457.0µ ± 0%   470.0µ ± 0%  +2.85% (p=0.000 n=20)   470.0µ ± 0%  +2.85% (p=0.000 n=20)    669.1µ ± 31%  +46.43% (p=0.003 n=20)
Memory/10000/lfvm-8        4.534m ± 1%   4.662m ± 0%  +2.81% (p=0.000 n=20)   4.662m ± 0%  +2.81% (p=0.000 n=20)    4.565m ±  0%   +0.67% (p=0.002 n=20)
Analysis/jumpdest/lfvm-8   425.4n ± 0%   437.6n ± 0%  +2.87% (p=0.000 n=20)   437.6n ± 0%  +2.87% (p=0.000 n=20)    433.5n ±  0%   +1.92% (p=0.000 n=20)
Analysis/stop/lfvm-8       425.8n ± 0%   437.4n ± 0%  +2.72% (p=0.000 n=20)   437.4n ± 0%  +2.72% (p=0.000 n=20)    432.9n ±  1%   +1.66% (p=0.000 n=20)
Analysis/push1/lfvm-8      427.1n ± 0%   439.9n ± 0%  +3.01% (p=0.000 n=20)   439.9n ± 0%  +3.01% (p=0.000 n=20)    433.1n ±  0%   +1.40% (p=0.000 n=20)
Analysis/push32/lfvm-8     422.7n ± 0%   437.8n ± 0%  +3.60% (p=0.000 n=20)   437.8n ± 0%  +3.60% (p=0.000 n=20)    433.1n ±  0%   +2.47% (p=0.000 n=20)
geomean                    12.52µ        12.81µ       +2.30%                  12.81µ       +2.30%                   13.67µ         +9.20%
```

The rightmost column is a version where the stack limits lookup has been changed from `[]{int, int}` to `[]{int16, int16}`, the change was done with the intention of reducing `satisfiesStackRequirements` execution time, but the outcome is difficult to explain. This change was removed from this PR.
![image](https://github.com/user-attachments/assets/4ea4cf4a-7968-49d7-906b-e429de1788fc)