@LuisPH3 2024-09-11
Benchmarks after `opcodeProperty` refactor:

```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/integration_test/interpreter
cpu: Intel(R) Xeon(R) Platinum 8481C CPU @ 2.70GHz
                         │   main_20   │         unified_opcodes_20         │
                         │   sec/op    │   sec/op     vs base               │
Inc/1/lfvm-8               1.672µ ± 0%   1.723µ ± 0%  +3.05% (p=0.000 n=20)
Inc/10/lfvm-8              1.670µ ± 0%   1.721µ ± 0%  +3.02% (p=0.000 n=20)
Fib/1/lfvm-8               1.434µ ± 0%   1.489µ ± 0%  +3.84% (p=0.000 n=20)
Fib/5/lfvm-8               9.033µ ± 0%   9.186µ ± 0%  +1.69% (p=0.000 n=20)
Fib/10/lfvm-8              94.89µ ± 0%   95.34µ ± 0%  +0.47% (p=0.000 n=20)
Fib/15/lfvm-8              1.047m ± 0%   1.051m ± 0%  +0.45% (p=0.000 n=20)
Fib/20/lfvm-8              11.58m ± 0%   11.67m ± 0%  +0.79% (p=0.000 n=20)
Sha3/1/lfvm-8              636.1n ± 0%   642.3n ± 0%  +0.97% (p=0.000 n=20)
Sha3/10/lfvm-8             1.937µ ± 0%   1.966µ ± 0%  +1.47% (p=0.000 n=20)
Sha3/100/lfvm-8            14.55µ ± 0%   14.66µ ± 0%  +0.76% (p=0.000 n=20)
Sha3/1000/lfvm-8           137.5µ ± 0%   139.6µ ± 0%  +1.56% (p=0.000 n=20)
Arith/1/lfvm-8             1.823µ ± 0%   1.895µ ± 0%  +3.95% (p=0.000 n=20)
Arith/10/lfvm-8            4.812µ ± 0%   4.927µ ± 0%  +2.39% (p=0.000 n=20)
Arith/100/lfvm-8           34.17µ ± 0%   34.27µ ± 0%  +0.31% (p=0.000 n=20)
Arith/280/lfvm-8           95.20µ ± 0%   95.46µ ± 0%  +0.27% (p=0.000 n=20)
Memory/1/lfvm-8            2.942µ ± 0%   3.008µ ± 0%  +2.24% (p=0.000 n=20)
Memory/10/lfvm-8           7.121µ ± 0%   7.270µ ± 1%  +2.10% (p=0.000 n=20)
Memory/100/lfvm-8          48.58µ ± 0%   49.16µ ± 2%  +1.18% (p=0.000 n=20)
Memory/1000/lfvm-8         456.7µ ± 0%   459.7µ ± 0%  +0.64% (p=0.000 n=20)
Memory/10000/lfvm-8        4.538m ± 0%   4.555m ± 0%  +0.38% (p=0.001 n=20)
Analysis/jumpdest/lfvm-8   426.9n ± 0%   429.8n ± 0%  +0.67% (p=0.000 n=20)
Analysis/stop/lfvm-8       426.4n ± 0%   430.6n ± 0%  +0.96% (p=0.000 n=20)
Analysis/push1/lfvm-8      427.7n ± 0%   431.9n ± 0%  +0.98% (p=0.000 n=20)
Analysis/push32/lfvm-8     427.0n ± 0%   434.2n ± 0%  +1.70% (p=0.000 n=20)
geomean                    12.48µ        12.67µ       +1.49%
```