@LuisPH3 2024-09-12
Updated benchmark after change to ```func (p *opCodePropertyMap[T]) get(op OpCode) T {```

```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/integration_test/interpreter
cpu: Intel(R) Xeon(R) Platinum 8481C CPU @ 2.70GHz
                         │   main_20   │         unified_opcodes_20         │
                         │   sec/op    │   sec/op     vs base               │
Inc/1/lfvm-8               1.672µ ± 0%   1.674µ ± 0%       ~ (p=0.149 n=20)
Inc/10/lfvm-8              1.670µ ± 0%   1.672µ ± 0%       ~ (p=0.941 n=20)
Fib/1/lfvm-8               1.434µ ± 0%   1.461µ ± 0%  +1.85% (p=0.000 n=20)
Fib/5/lfvm-8               9.033µ ± 0%   9.023µ ± 0%  -0.12% (p=0.007 n=20)
Fib/10/lfvm-8              94.89µ ± 0%   94.00µ ± 0%  -0.94% (p=0.000 n=20)
Fib/15/lfvm-8              1.047m ± 0%   1.035m ± 0%  -1.09% (p=0.000 n=20)
Fib/20/lfvm-8              11.58m ± 0%   11.45m ± 0%  -1.13% (p=0.000 n=20)
Sha3/1/lfvm-8              636.1n ± 0%   633.6n ± 0%  -0.41% (p=0.000 n=20)
Sha3/10/lfvm-8             1.937µ ± 0%   1.957µ ± 0%  +1.03% (p=0.000 n=20)
Sha3/100/lfvm-8            14.55µ ± 0%   14.50µ ± 0%  -0.38% (p=0.000 n=20)
Sha3/1000/lfvm-8           137.5µ ± 0%   137.6µ ± 0%  +0.09% (p=0.011 n=20)
Arith/1/lfvm-8             1.823µ ± 0%   1.824µ ± 0%       ~ (p=0.773 n=20)
Arith/10/lfvm-8            4.812µ ± 0%   4.834µ ± 0%  +0.47% (p=0.000 n=20)
Arith/100/lfvm-8           34.17µ ± 0%   34.20µ ± 0%  +0.09% (p=0.038 n=20)
Arith/280/lfvm-8           95.20µ ± 0%   95.29µ ± 0%       ~ (p=0.071 n=20)
Memory/1/lfvm-8            2.942µ ± 0%   2.948µ ± 0%  +0.22% (p=0.043 n=20)
Memory/10/lfvm-8           7.121µ ± 0%   7.163µ ± 0%  +0.59% (p=0.000 n=20)
Memory/100/lfvm-8          48.58µ ± 0%   48.48µ ± 0%  -0.21% (p=0.002 n=20)
Memory/1000/lfvm-8         456.7µ ± 0%   455.4µ ± 0%  -0.30% (p=0.000 n=20)
Memory/10000/lfvm-8        4.538m ± 0%   4.511m ± 0%  -0.58% (p=0.000 n=20)
Analysis/jumpdest/lfvm-8   426.9n ± 0%   425.8n ± 0%       ~ (p=0.109 n=20)
Analysis/stop/lfvm-8       426.4n ± 0%   427.6n ± 0%       ~ (p=0.171 n=20)
Analysis/push1/lfvm-8      427.7n ± 0%   428.3n ± 0%       ~ (p=0.713 n=20)
Analysis/push32/lfvm-8     427.0n ± 0%   427.3n ± 0%       ~ (p=0.457 n=20)
geomean                    12.48µ        12.48µ       -0.01%
```